+++ frontend/src/container/GridCardLayout/GridCard/index.tsx
@@ -30,6 +30,8 @@ function GridCardGraph({
 	threshold,
 	variables,
 	version,
+	onClickHandler,
+	onDragSelect,
 }: GridCardGraphProps): JSX.Element {
 	const dispatch = useDispatch();
 	const [errorMessage, setErrorMessage] = useState<string>();
@@ -178,6 +180,8 @@ function GridCardGraph({
 					headerMenuList={menuList}
 					isFetchingResponse={queryResponse.isFetching}
 					setRequestData={setRequestData}
+					onClickHandler={onClickHandler}
+					onDragSelect={onDragSelect}
 				/>
 			)}
 		</div>
