+++ frontend/src/container/MetricsApplication/Tabs/Overview/TopLevelOperations.tsx
@@ -28,8 +28,6 @@ function TopLevelOperation({
 			) : (
 				<GraphContainer>
 					<Graph
-						fillSpans={false}
-						name={name}
 						widget={widget}
 						onClickHandler={handleGraphClick(opName)}
 						onDragSelect={onDragSelect}
