+++ frontend/src/container/PanelWrapper/UplotPanelWrapper.tsx
@@ -1,22 +1,22 @@
 import { ToggleGraphProps } from 'components/Graph/types';
 import Uplot from 'components/Uplot';
-import { QueryParams } from 'constants/query';
+// import { QueryParams } from 'constants/query';
 import { PANEL_TYPES } from 'constants/queryBuilder';
 import GraphManager from 'container/GridCardLayout/GridCard/FullView/GraphManager';
 import { getLocalStorageGraphVisibilityState } from 'container/GridCardLayout/GridCard/utils';
 import { useIsDarkMode } from 'hooks/useDarkMode';
 import { useResizeObserver } from 'hooks/useDimensions';
-import useUrlQuery from 'hooks/useUrlQuery';
-import GetMinMax from 'lib/getMinMax';
-import history from 'lib/history';
+// import useUrlQuery from 'hooks/useUrlQuery';
+// import GetMinMax from 'lib/getMinMax';
+// import history from 'lib/history';
 import { getUPlotChartOptions } from 'lib/uPlotLib/getUplotChartOptions';
 import { getUPlotChartData } from 'lib/uPlotLib/utils/getUplotChartData';
 import _noop from 'lodash-es/noop';
 import { useDashboard } from 'providers/Dashboard/Dashboard';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { useLocation } from 'react-router-dom';
-import { UpdateTimeInterval } from 'store/actions';
+import { useEffect, useMemo, useRef, useState } from 'react';
+// import { useDispatch } from 'react-redux';
+// import { useLocation } from 'react-router-dom';
+// import { UpdateTimeInterval } from 'store/actions';
 import { getSortedSeriesData } from 'utils/getSortedSeriesData';
 import { getTimeRange } from 'utils/getTimeRange';
 
@@ -29,14 +29,16 @@ function UplotPanelWrapper({
 	setGraphVisibility,
 	graphVisibility,
 	onToggleModelHandler,
+	onClickHandler,
+	onDragSelect,
 }: PanelWrapperProps): JSX.Element {
-	const dispatch = useDispatch();
+	// const dispatch = useDispatch();
 	const { toScrollWidgetId, setToScrollWidgetId } = useDashboard();
 	const isDarkMode = useIsDarkMode();
 	const lineChartRef = useRef<ToggleGraphProps>();
 	const graphRef = useRef<HTMLDivElement>(null);
-	const urlQuery = useUrlQuery();
-	const location = useLocation();
+	// const urlQuery = useUrlQuery();
+	// const location = useLocation();
 	const [minTimeScale, setMinTimeScale] = useState<number>();
 	const [maxTimeScale, setMaxTimeScale] = useState<number>();
 
@@ -60,27 +62,27 @@ function UplotPanelWrapper({
 
 	const containerDimensions = useResizeObserver(graphRef);
 
-	const onDragSelect = useCallback(
-		(start: number, end: number): void => {
-			const startTimestamp = Math.trunc(start);
-			const endTimestamp = Math.trunc(end);
+	// const onDragSelect = useCallback(
+	// 	(start: number, end: number): void => {
+	// 		const startTimestamp = Math.trunc(start);
+	// 		const endTimestamp = Math.trunc(end);
 
-			if (startTimestamp !== endTimestamp) {
-				dispatch(UpdateTimeInterval('custom', [startTimestamp, endTimestamp]));
-			}
+	// 		if (startTimestamp !== endTimestamp) {
+	// 			dispatch(UpdateTimeInterval('custom', [startTimestamp, endTimestamp]));
+	// 		}
 
-			const { maxTime, minTime } = GetMinMax('custom', [
-				startTimestamp,
-				endTimestamp,
-			]);
+	// 		const { maxTime, minTime } = GetMinMax('custom', [
+	// 			startTimestamp,
+	// 			endTimestamp,
+	// 		]);
 
-			urlQuery.set(QueryParams.startTime, minTime.toString());
-			urlQuery.set(QueryParams.endTime, maxTime.toString());
-			const generatedUrl = `${location.pathname}?${urlQuery.toString()}`;
-			history.push(generatedUrl);
-		},
-		[dispatch, location.pathname, urlQuery],
-	);
+	// 		urlQuery.set(QueryParams.startTime, minTime.toString());
+	// 		urlQuery.set(QueryParams.endTime, maxTime.toString());
+	// 		const generatedUrl = `${location.pathname}?${urlQuery.toString()}`;
+	// 		history.push(generatedUrl);
+	// 	},
+	// 	[dispatch, location.pathname, urlQuery],
+	// );
 
 	useEffect(() => {
 		const {
@@ -116,7 +118,7 @@ function UplotPanelWrapper({
 				isDarkMode,
 				onDragSelect,
 				yAxisUnit: widget?.yAxisUnit,
-				onClickHandler: _noop,
+				onClickHandler: onClickHandler || _noop,
 				thresholds: widget.thresholds,
 				minTimeScale,
 				maxTimeScale,
@@ -137,6 +139,7 @@ function UplotPanelWrapper({
 			containerDimensions,
 			isDarkMode,
 			onDragSelect,
+			onClickHandler,
 			minTimeScale,
 			maxTimeScale,
 			graphVisibility,
